# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Nocalhost server api",
        "title": "Nocalhost docs api",
        "contact": {
            "name": "wangwei",
            "url": "nocalhost.coding.net"
        },
        "license": {},
        "version": "1.0"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
...
        "/v1/cluster/{id}/dev_space": {
            "get": {
                "description": "Cluster entrance to obtain cluster development environment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cluster"
                ],
                "summary": "Cluster dev space list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "kubeconfig",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.ClusterUserModel"
                        }
                    }
                }
            }
        },
...
        "/v1/dev_space": {
            "post": {
                "description": "Create dev space",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DevSpace"
                ],
                "summary": "Create dev space",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "dev space info",
                        "name": "CreateAppRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/cluster_user.ClusterUserCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.ClusterUserModel"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "cluster_user.ClusterUserCreateRequest": {
            "type": "object",
            "required": [
                "cluster_id",
                "user_id"
            ],
            "properties": {
                "application_id": {
                    "type": "integer"
                },
                "base_dev_space_id": {
                    "type": "integer"
                },
                "cluster_admin": {
                    "type": "integer"
                },
                "cluster_id": {
                    "type": "integer"
                },
                "cpu": {
                    "type": "integer"
                },
                "dev_space_type": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_base_space": {
                    "type": "boolean"
                },
                "memory": {
                    "type": "integer"
                },
                "mesh_dev_info": {
                    "type": "object",
                    "$ref": "#/definitions/setupcluster.MeshDevInfo"
                },
                "namespace": {
                    "type": "string"
                },
                "protected": {
                    "type": "boolean"
                },
                "space_name": {
                    "type": "string"
                },
                "space_resource_limit": {
                    "type": "object",
                    "$ref": "#/definitions/cluster_user.SpaceResourceLimit"
                },
                "user_id": {
                    "type": "integer"
                },
                "virtual_cluster": {
                    "type": "object",
                    "$ref": "#/definitions/model.VirtualClusterInfo"
                }
            }
        },
...
        "cluster_user.SpaceResourceLimit": {
            "type": "object",
            "properties": {
                "container_ephemeral_storage": {
                    "type": "string"
                },
                "container_limits_cpu": {
                    "type": "string"
                },
                "container_limits_mem": {
                    "type": "string"
                },
                "container_req_cpu": {
                    "type": "string"
                },
                "container_req_mem": {
                    "type": "string"
                },
                "space_ephemeral_storage": {
                    "type": "string"
                },
                "space_lb_count": {
                    "type": "string"
                },
                "space_limits_cpu": {
                    "type": "string"
                },
                "space_limits_mem": {
                    "type": "string"
                },
                "space_pvc_count": {
                    "type": "string"
                },
                "space_req_cpu": {
                    "type": "string"
                },
                "space_req_mem": {
                    "type": "string"
                },
                "space_storage_capacity": {
                    "type": "string"
                }
            }
        },
...
        "model.ClusterUserModel": {
            "type": "object",
            "properties": {
                "application_id": {
                    "description": "Deprecated",
                    "type": "integer"
                },
                "base_dev_space_id": {
                    "type": "integer"
                },
                "cluster_admin": {
                    "type": "integer"
                },
                "cluster_id": {
                    "type": "integer"
                },
                "cpu": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "dev_space_type": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_base_space": {
                    "type": "boolean"
                },
                "kubeconfig": {
                    "type": "string"
                },
                "memory": {
                    "type": "integer"
                },
                "namespace": {
                    "type": "string"
                },
                "protected": {
                    "type": "boolean"
                },
                "space_name": {
                    "type": "string"
                },
                "space_resource_limit": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "trace_header": {
                    "type": "object",
                    "$ref": "#/definitions/model.Header"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "model.Header": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "model.VirtualClusterInfo": {
            "type": "object",
            "properties": {
                "events": {
                    "type": "string"
                },
                "service_type": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "values": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "setupcluster.MeshDevApp": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "workloads": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/setupcluster.MeshDevWorkload"
                    }
                }
            }
        },
        "setupcluster.MeshDevInfo": {
            "type": "object",
            "properties": {
                "apps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/setupcluster.MeshDevApp"
                    }
                },
                "header": {
                    "type": "object",
                    "$ref": "#/definitions/model.Header"
                },
                "namespace": {
                    "type": "string"
                },
                "resources": {
                    "type": "object",
                    "$ref": "#/definitions/setupcluster.meshDevResources"
                },
                "rollback": {
                    "type": "object",
                    "$ref": "#/definitions/setupcluster.rollback"
                }
            }
        },
        "setupcluster.MeshDevWorkload": {
            "type": "object",
            "properties": {
                "kind": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "setupcluster.meshDevResources": {
            "type": "object",
            "properties": {
                "delete": {
                    "type": "string"
                },
                "install": {
                    "type": "string"
                }
            }
        },
        "setupcluster.rollback": {
            "type": "object",
            "properties": {
                "header": {
                    "type": "object",
                    "$ref": "#/definitions/setupcluster.rollbackHeader"
                }
            }
        },
        "setupcluster.rollbackHeader": {
            "type": "object",
            "properties": {
                "add": {
                    "type": "object"
                },
                "update": {
                    "type": "object"
                }
            }
        },
...
    }
}